cleaning up my backend code
