+++ todo.txt
@@ -9,6 +9,13 @@ TO DO:
         - Currently when someone purchases multiple things the backend is not updating
         - Add parent details into the details that go into the camps
         - Day 2 is not working (think it could be a python issue)
+        - update SETCOACHTIMES so that it doesnt add a new coach everytime, it just adds onto their current times
 
     General ##########################################
         - Update the render so that new env variable for FRONT_URL
+        - make new property "archived" as a boolean for camps
+
+
+######## What to update from dev to production  #####
+- URL on front end (to right backend)
+
